campaign_engine/requirements.txt
kafka-python==2.0.2
google-generativeai==0.7.2
tenacity==8.2.3
requests==2.31.0
